fix: use torch_dtype in from_pretrained for transformers>=4.45 compat
dtype= is a newer kwarg alias for AutoModelForCausalLM.from_pretrained
and may be silently ignored on older-but-in-range transformers installs
near the declared >=4.45 floor, causing the user's selected precision
to be dropped in favor of the default. torch_dtype= is recognized
across the full supported range.

+++ comfyui_realtime/providers/transformers_llm.py
@@ -100,7 +100,7 @@ class TransformersLLMProvider:
             )
 
         self._model = AutoModelForCausalLM.from_pretrained(
-            model_path, dtype=resolved_dtype, trust_remote_code=trust_remote_code
+            model_path, torch_dtype=resolved_dtype, trust_remote_code=trust_remote_code
         )
         self._model.to(resolved_device)
         self._model.eval()
